requirements.txt
Django>=1.8,<=1.10
requests>=2.0
requirements_parser==0.1.0
semantic_version>=2.5
